rename job_status
