+++ dpdispatcher/machines/distributed_shell.py
@@ -1,5 +1,5 @@
 from dpdispatcher.dlog import dlog
-from dpdispatcher.utils.JobStatus import JobStatus
+from dpdispatcher.utils.job_status import JobStatus
 from dpdispatcher.machine import Machine
 from dpdispatcher.utils.utils import (
     customized_script_header_template,
