+++ dpdispatcher/machines/lsf.py
@@ -4,7 +4,7 @@ from typing import List
 from dargs import Argument
 
 from dpdispatcher.dlog import dlog
-from dpdispatcher.utils.JobStatus import JobStatus
+from dpdispatcher.utils.job_status import JobStatus
 from dpdispatcher.machine import Machine
 from dpdispatcher.utils.utils import RetrySignal, customized_script_header_template, retry
