+++ dpdispatcher/machines/openapi.py
@@ -15,7 +15,7 @@ else:
     found_bohriumsdk = True
 
 from dpdispatcher.dlog import dlog
-from dpdispatcher.utils.JobStatus import JobStatus
+from dpdispatcher.utils.job_status import JobStatus
 from dpdispatcher.machine import Machine
 
 shell_script_header_template = """
