+++ dpdispatcher/machines/slurm.py
@@ -6,7 +6,7 @@ from typing import List
 from dargs import Argument
 
 from dpdispatcher.dlog import dlog
-from dpdispatcher.utils.JobStatus import JobStatus
+from dpdispatcher.utils.job_status import JobStatus
 from dpdispatcher.machine import Machine, script_command_template
 from dpdispatcher.utils.utils import RetrySignal, customized_script_header_template, retry
